+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # MeshAdminUI Changelog
 
-## Version 0.8.2 (tbd)
+## Version 0.8.2 (upcoming)
+* Nodes in the "Select Node..." dialog are now sorted by their display name. (#15)
+* The "Select Node..." dialog now remembers the last position it was opened. (#12)
 * Add multi binary upload dialog. Users can now upload multiple files at once by clicking the button next to the create node button.
 
 ## Version 0.8.1 (2017-11-29)